back/requirements/base.txt
annotated-types==0.5.0
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
Automat==20.2.0
# amqp==5.0.6
amqp==5.1.1
appdirs==1.4.4
asgiref==3.4.1
astroid==2.5
autopep8==1.5.7
attrs==21.4.0
autobahn==22.3.2
Babel==2.9.1
billiard==3.6.4.0
botocore==1.20.43
branca==0.4.2
# celery==5.1.1
celery==5.2.2
# certifi==2021.10.8
certifi==2022.12.07
cffi==1.14.6
channels==3.0.4
constantly==15.1.0
cryptocode==0.1
# cryptography==36.0.2
# cryptography==39.0.1
cryptography==38.0.4
chardet==3.0.4
# click==7.1.2
click==8.1.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.25.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
daphne==3.0.2
distlib==0.3.1
dj-database-url==0.5.0
dj-places==4.0.0
# Django==3.2.7
Django==3.2.10
django-admin-tools==0.9.1
django-appconf==1.0.4
django-celery-beat==2.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-ckeditor==6.1.0
django-cacheops==6.1
django-js-asset==1.2.2
django-debug-toolbar==3.2.2
django-debug-toolbar-template-profiler==2.0.2
django-filter==2.4.0
django-redis==5.0.0
django-taggit==1.4.0
django-timezone-field==4.2.3
django-yamlfield==1.0.3
djangorestframework==3.12.4
django-rest-swagger==2.2.0
django-phone-field==1.8.1
et-xmlfile==1.1.0
filelock==3.0.12
flake8==3.9.2
folium==0.12.1
futures==3.1.1
frozenlist==1.3.3
funcy==1.18
geographiclib==1.52
geoip2==2.9.0
geopy==2.2.0
googletrans==3.0.0
goslate==1.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
# httpx==0.23.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
isort==5.7.0
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
# joblib==1.0.1
joblib==1.2.0
pyasn1==0.4.8
pycryptodomex==3.16.0
# kombu==5.1.0
kombu==5.2.4
lazy-object-proxy==1.5.2
loguru==0.6.0
lml==0.1.0
# lxml==4.6.4
lxml==4.6.5
MarkupSafe==2.0.1
magic-filter==1.0.9
maxminddb==2.0.3
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
# nltk==3.6.2
nltk==3.6.6
# numpy==1.21.0
# numpy==1.22
# oauthlib==3.1.1
oauthlib==3.2.2
openpyxl==3.0.7
openapi-codec==1.3.2
paho-mqtt==1.6.1
pandas==1.3.0
pathspec==0.8.1
pep8==1.7.1
# Pillow==8.3.1
Pillow==8.3.2
prompt-toolkit==3.0.19
psycopg2-binary==2.9.5
pycodestyle==2.7.0
pycparser==2.20
pydantic==2.3.0
pydantic-core==2.6.3
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
pyflakes==2.3.1
pyTelegramBotAPI==4.7.1
# PyJWT==2.1.0
PyJWT==2.4.0
pylint==2.7.1
pylint-plugin-utils==0.6
python-crontab==2.6.0
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyasn1-modules==0.2.8
pyopenssl==22.0.0
priority==1.3.0
# redis==3.5.3
redis==4.3.6
# regex==2021.4.4
regex==2023.3.23
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
s3transfer==0.3.6
six==1.15.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
simplejson==3.19.1
# sqlparse==0.4.1
sqlparse==0.4.2
service-identity==21.1.0
textblob==0.15.3
texttable==1.6.4
toml==0.10.2
transliterate==1.10.2
tqdm==4.62.0
# twisted==22.4.0
twisted==22.10.0
txaio==22.2.1
typing-extensions==4.7.1
# untangle==1.1.1
untangle==1.2.1
# urllib3==1.26.4
uritemplate==4.1.1
urllib3==1.26.5
uvicorn==0.24.0.post1
vine==5.0.0
virtualenv==20.4.3
wcwidth==0.2.5
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
yarl==1.8.1
zope.interface==5.4.0
# git+https://github.com/ilmix007/fapi@main
